# TranspozUI/src/app/services/models/envelope.ts
/* tslint:disable */
/* eslint-disable */
export interface Envelope {
  area?: number;
  diameter?: number;
  height?: number;
  maxX?: number;
  maxY?: number;
  minX?: number;
  minY?: number;
  null?: boolean;
  width?: number;
}