# TranspozUI/src/app/services/models/precision-model.ts
/* tslint:disable */
/* eslint-disable */
import { Type } from '../models/type';
export interface PrecisionModel {
  floating?: boolean;
  maximumSignificantDigits?: number;
  offsetX?: number;
  offsetY?: number;
  scale?: number;
  type?: Type;
}