# TranspozUI/src/app/services/models/vehicle.ts
/* tslint:disable */
/* eslint-disable */
export interface Vehicle {
  capacity?: number;
  color?: string;
  id?: number;
  licensePlate?: string;
  make?: string;
  model?: string;
  type?: 'CAR' | 'VAN' | 'TRUCK';
  year?: number;
}